Make src/crypto and src/controller compile with -Wconversion (#2810)
* Use uint16_t for our various header sizes in bytes

PacketBuffer uses uint16_t for all of its sizes already.  It's easier
to tell that we're not overflowing the buffer if we know our header
sizes also fit in uint16_t.

* Consistently use uint16_t for key ids.

This prevents -Wconversion warnings when shuffling key ids around.

* Make src/crypto and src/controller compile with -Werror=conversion

+++ src/crypto/tests/PBKDF2_SHA256_test_vectors.h
@@ -34,7 +34,7 @@ struct pbkdf2_test_vector
     const uint8_t * salt;
     size_t slen;
     unsigned int iter;
-    size_t key_len;
+    uint32_t key_len;
     const uint8_t * key;
     unsigned tcId;
     int result;
